scraper/requirements.txt
# Core dependencies for the Management & Marketing journal scraper

# HTTP requests library
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional: Progress bars for CLI usage
# tqdm>=4.66.0

# Optional: Data validation
# pydantic>=2.0.0
